feat: 로그인 기능까지 추가

+++ backend/src/main/java/com/example/SMU_WordMaster/service/WordService.java
@@ -78,7 +78,10 @@ public class WordService {
     public void deleteWord(List<String> wordList) {
         for (String s : wordList) {
             Words wordEntity = utils.getWordsEntity(s);
-            wordsRepository.deleteById(wordEntity.getWordId());
+            String spelling = wordEntity.getSpelling();
+            // 중복 저장 방지
+            if (!wordsRepository.existsBySpelling(spelling))
+                wordsRepository.deleteById(wordEntity.getWordId());
         }
     }
 }
